wav streaming, audio cleanup
there’s a bug in the timing / byte tracking streams (of both wav and
ogg) and when skipping to 75% of the wav it makes white noise for some
reason (might be related). Need to think of a more effective way of
tracking the byte frames.

fixing crashes in audio context cleanup, and sound destroy throwing
errors

adding wav streaming chunking and seeking etc