Refactoring IO and adding bindings for an IO file
Use Assets, get_bytes etc. These are for low level helpers -

There are haxe files yes, but this is bound directly to the cross
platform lumen IO such that 1) it matches fread/rwops and 2) handles
any apk crap, etc.

Also, this is untested. Tests will be made in the great Testening of
the lumen code.