Large scale refactoring to accommodate html5 builds, ok if tests fail right now but shouldn't
Note this is half way done. Assets and Audio need refactoring to
platform specific folder and i am not done with the naming conventions,
*System might be changed out etc.

This includes unifying all platform separation in one place, and
pointing the classes to those.
Includes refactoring the GL code to be clearer and less import-y
Includes removing all haxe3 defines from the utils functions