utils; add ByteArrayReadOptions for async/binary and future flags to be added. This defaults to UTF-8 encoding for non-binary assets on web to correctly transfer. close #61
In future we could optionize the mime type and char encoding but the
idea is to reduce parity across platforms, not make a general purpose
xmlhttprequest wrapper. In these cases you might best just use the js
api directly and create a byte array from it in the same fashion.